Remove info about custom attributes not prefixed with `data-`.
For ng-l20n using a `l20n` attribute instead of `data-l20n` is discouraged (though
it would work) as using any custom attributes without a dash in their name is
not future-proof as future HTML version might define such attributes.

As for `l20n.js` itself, the `l10n-id` attribute is not supported and using it will
make the code break.

Because of all that, this commit removes information about using custom
attributes without prefixing them with `data-`.