Fix: delay setting $rootScope[localeProp] until locale is fully switched
The locale needs to be set on the `ready` event since `requestLocales`
can be asynchronous. We can't just use `documentL10n.ready`, though
as context once marked ready is never unmarked as such. :-(
Thus, we have to register the handler before the `requestLocale`
invocation; otherwise we have no way of knowing if the event alredy
fired.