Add empty attribute checking
Checking if the attribute is truthy prevents from passing an empty translation
key to l20n. If an empty key is used on any node then l20n will translate
neither that node nor any of its following nodes in the entire document. In
the worst case scenario this may lead to a permanent l20n failure if the empty
key occurs at the very beginning of the document.

Closes gh-9